+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/ProductListDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.dto
+package woowacourse.shopping.data.remote.dto.response
 
 data class ProductListDto(
     val products: List<ProductDto>,
