+++ app/src/main/java/woowacourse/shopping/data/remote/dto/response/ShoppingCartDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.dto
+package woowacourse.shopping.data.remote.dto.response
 
 import woowacourse.shopping.domain.model.Product
